ajout de sécurité pour les fonctions

+++ site_recette_avec_mvc/Controleurs/ControleurAdmin.php
@@ -5,32 +5,36 @@ final class ControleurAdmin
     private $model;
     public function defautAction()
     {
-        $this->model = new Admin();
-        $_SESSION['users_list'] = $this->model->getUsers();
-        Vue::montrer('admin/voir');
+        if (isset($_SESSION['userAdmin']) and $_SESSION['userAdmin'] == true){
+            $this->model = new Admin();
+            $_SESSION['users_list'] = $this->model->getUsers();
+            Vue::montrer('admin/voir');
+        }
     }
 
     public function supprimerUtilisateurAction() {
-        // get id
-        if(isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on')   
-            $url = "https://";   
-        else  
-            $url = "http://";   
-        // Append the host(domain name, ip) to the URL.   
-        $url.= $_SERVER['HTTP_HOST'];   
-        // Append the requested resource location to the URL   
-        $url.= $_SERVER['REQUEST_URI'];
-        // id = url.split("=")[1];
-        $_SESSION["id_user"] = explode("=", $url)[1];
+        if (isset($_SESSION['userAdmin']) and $_SESSION['userAdmin'] == true){
+            // get id
+            if(isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on')   
+                $url = "https://";   
+            else  
+                $url = "http://";   
+            // Append the host(domain name, ip) to the URL.   
+            $url.= $_SERVER['HTTP_HOST'];   
+            // Append the requested resource location to the URL   
+            $url.= $_SERVER['REQUEST_URI'];
+            // id = url.split("=")[1];
+            $_SESSION["id_user"] = explode("=", $url)[1];
 
-        // instancier le modèle Admin
-        $this->model = new Admin();
-    
-        // utiliser la méthode deleteUsers pour supprimer un utilisateur
-        $this->model->deleteUsers($_SESSION["id_user"]);
-    
-        // rediriger l'utilisateur vers la liste des utilisateurs
-        header('location: /admin');
+            // instancier le modèle Admin
+            $this->model = new Admin();
+        
+            // utiliser la méthode deleteUsers pour supprimer un utilisateur
+            $this->model->deleteUsers($_SESSION["id_user"]);
+        
+            // rediriger l'utilisateur vers la liste des utilisateurs
+            header('location: /admin');
+        }
     }
 
     public function checkAdminAction()
@@ -55,17 +59,19 @@ final class ControleurAdmin
         }
     }
     public function supprimerRecetteAction() {
-        // récupérer l'id de la recette à supprimer
-        $id = $_GET['id'];
-        var_dump($_GET['id']);
-        echo $_GET['id'];
-        // instancier le modèle de recette
-        $recetteModel = new Admin();
-    
-        // utiliser la méthode deleteRecette pour supprimer la recette
-        $recetteModel->deleteRecette($id);
-    
-        // rediriger l'utilisateur vers la liste des recettes
-        //header('location: /liste-recettes.php');
+        if (isset($_SESSION['userAdmin']) and $_SESSION['userAdmin'] == true){
+            // récupérer l'id de la recette à supprimer
+            $id = $_GET['id'];
+            var_dump($_GET['id']);
+            echo $_GET['id'];
+            // instancier le modèle de recette
+            $recetteModel = new Admin();
+        
+            // utiliser la méthode deleteRecette pour supprimer la recette
+            $recetteModel->deleteRecette($id);
+        
+            // rediriger l'utilisateur vers la liste des recettes
+            //header('location: /liste-recettes.php');
+        }
     }
 }
